*open "must-gather" folder "ocs_ci/ocs/must_gather" *split for 3 test cases (type of logs: ceph,oc output, noobaa)
